requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy<2
opencv-python==4.8.1.78
tqdm
tensorboard
tensorboardX
scikit-image
medpy
pillow
scipy
PyYAML
lmdb
matplotlib
datasets==4.4.1
